# [compiler] better type inference for switch cond var