--- a/compiler/code-gen/vertex-compiler.cpp
+++ b/compiler/code-gen/vertex-compiler.cpp
@@ -1108,7 +1108,7 @@ void compile_switch_str(VertexAdaptor<op_switch> root, CodeGenerator &W) {
   W << temp_var_strval_of_condition << " = f$strval (" << root->condition() << ");" << NL;
   W << temp_var_matched_with_a_case << " = false;" << NL;
 
-  W << "switch (" << temp_var_strval_of_condition << ".as_string().hash()) " << BEGIN;
+  W << "switch (" << temp_var_strval_of_condition << ".hash()) " << BEGIN;
   for (const auto &one_case : cases) {
     if (one_case.is_default) {
       W << "default:" << NL;
@@ -1124,7 +1124,7 @@ void compile_switch_str(VertexAdaptor<op_switch> root, CodeGenerator &W) {
       }
       W << "if (!" << temp_var_matched_with_a_case << ") " << BEGIN;
       {
-        W << "if (!equals (" << temp_var_strval_of_condition << ".as_string(), " << one_case.expr << ")) " << BEGIN;
+        W << "if (!equals (" << temp_var_strval_of_condition << ", " << one_case.expr << ")) " << BEGIN;
 
         if (one_case.next) {
           W << "goto " << one_case.next->goto_name << ";" << NL;
@@ -1231,9 +1231,6 @@ void compile_switch_var(VertexAdaptor<op_switch> root, CodeGenerator &W) {
 
 void compile_switch(VertexAdaptor<op_switch> root, CodeGenerator &W) {
   kphp_assert(root->condition_on_switch()->type() == op_var && root->matched_with_one_case()->type() == op_var);
-  bool all_cases_are_ints = true;
-  int num_const_string_cases = 0;
-  int num_value_cases = 0;
   bool has_default = false;
 
   for (auto one_case : root->cases()) {
@@ -1242,22 +1239,11 @@ void compile_switch(VertexAdaptor<op_switch> root, CodeGenerator &W) {
       has_default = true;
       continue;
     }
-
-    auto val = GenTree::get_actual_value(one_case.as<op_case>()->expr());
-    all_cases_are_ints    &= is_const_int(val);
-    num_value_cases++;
-    if (auto as_string = val.try_as<op_string>()) {
-      // PHP would use a numerical comparison for strings that look like a number,
-      // we shouldn't rewrite these switches as a string-only switch
-      if (!php_is_numeric(as_string->str_val.data())) {
-        num_const_string_cases++;
-      }
-    }
   }
 
-  if (num_const_string_cases == num_value_cases) {
+  if (root->kind == SwitchKind::StringSwitch) {
     compile_switch_str(root, W);
-  } else if (all_cases_are_ints) {
+  } else if (root->kind == SwitchKind::IntSwitch) {
     compile_switch_int(root, W);
   } else {
     compile_switch_var(root, W);