--- a/compiler/vertex-desc.json
+++ b/compiler/vertex-desc.json
@@ -750,6 +750,11 @@
     "props": {
       "str": "switch"
     },
+    "extra_fields": {
+      "kind": {
+        "type": "SwitchKind"
+      }
+    },
     "ranges": {
       "cases": [
         3,