Instead of producing a mixed-typed variable for every switch statement, try to find a more suitable type.

Most of the switches are optimized to int and string switches, so in those cases having non-mixed var is very beneficial.

Even for expr (var) switches we can get a measurable improvement if all cases and condition expr itself evaluate to some properly-typed value, like int or string.

Here is a non-compherensive list of why mixed-typed vars are not optimal:

* Expensive destructor, `mixed::~mixed` shows up in the CPU profiles
* The need to do `mixed::as_string()` many times inside optimized string switches
* mixed-typed overloadings are used in `eq2()` calls

The implementation used in this PR tries to avoid excessive work from the compiler. If it can be assumed to be a int/string switch, we won't do any extra work. For other switch statements we'll find an LCA type of condition expr and all expression cases results. It could still be mixed, but sometimes it's inferred to something like int and improves performance significantly.

	name                      old time/op  new time/op  delta
	MultiSwitch::MultiString  83.0ns ± 0%  59.3ns ± 1%  -28.55%  (p=0.000 n=8+10)
	MultiSwitch::MultiInt     23.3ns ± 3%   7.0ns ± 0%  -69.96%  (p=0.000 n=10+10)